[clients] Migrate ToolbeltConfig
